requirements.txt
tensorflow==2.15.1
keras==2.15.0
keras-core==0.1.7
keras-cv==0.9.0
numpy
Pillow
streamlit
